requirements.txt
pandas==1.5.2
Pillow==9.5.0
scikit_learn==1.2.0
streamlit==1.17.0
